refactor: move dummy-data into src folder
